requirement.txt
altgraph==0.17.3
llvmlite==0.39.1
numba==0.56.4
numpy==1.23.5
pefile==2023.2.7
Pillow==9.4.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt5Designer==5.14.1
pywin32-ctypes==0.2.0
xmltodict==0.13.0
